+++ lib/gn_list_resolver.rb
@@ -8,6 +8,7 @@ require "logger"
 require "logger/colors"
 require "pp"
 require "biodiversity"
+require "concurrent"
 require "gn_uuid"
 require "graphql/client"
 require "graphql/client/http"
@@ -20,6 +21,7 @@ require "gn_list_resolver/writer"
 require "gn_list_resolver/collector"
 require "gn_list_resolver/column_collector"
 require "gn_list_resolver/sci_name_collector"
+require "gn_list_resolver/resolver_job"
 require "gn_list_resolver/resolver"
 require "gn_list_resolver/result_processor"
 require "gn_list_resolver/stats"
@@ -37,7 +39,7 @@ module GnListResolver
       reader = create_reader(input_io, opts)
       data = block_given? ? reader.read(&Proc.new) : reader.read
       writer = create_writer(reader, output_io, opts)
-      resolver = create_resolver(writer, opts)
+      resolver = Resolver.new(writer, opts)
       block_given? ? resolver.resolve(data, &Proc.new) : resolver.resolve(data)
       logger.warn(resolver.stats.stats.pretty_inspect) if opts[:debug]
       resolver.stats
@@ -61,12 +63,17 @@ module GnListResolver
       end
     end
 
-    private
-
-    def create_resolver(writer, opts)
-      Resolver.new(writer, opts.data_source_id,
-                   opts.stats, opts.with_classification)
+    def opts_struct(opts)
+      threads = opts[:threads].to_i
+      opts[:threads] = threads.between?(1, 10) ? threads : 2
+      with_classification = opts[:with_classification] ? true : false
+      opts[:with_classification] = with_classification
+      data_source_id = opts[:data_source_id].to_i
+      opts[:data_source_id] = data_source_id.zero? ? 1 : data_source_id
+      OpenStruct.new({ stats: Stats.new, alt_headers: [] }.merge(opts))
     end
+
+    private
 
     def create_writer(reader, output_io, opts)
       Writer.new(output_io, reader.original_fields,
@@ -76,10 +83,6 @@ module GnListResolver
     def create_reader(input_io, opts)
       Reader.new(input_io, input_name(opts.input),
                  opts.skip_original, opts.alt_headers, opts.stats)
-    end
-
-    def opts_struct(opts)
-      OpenStruct.new({ stats: Stats.new, alt_headers: [] }.merge(opts))
     end
 
     def io(input, output)