+++ lib/gn_list_resolver/resolver_job.rb
@@ -1,0 +1,39 @@
+# frozen_string_literal: true
+
+module GnListResolver
+  # Remote resolution for parallel jobs
+  class ResolverJob
+    GRAPHQL = GnGraphQL.new
+    QUERY = GRAPHQL.client.parse(GRAPHQL.query)
+    def initialize(names, batch_data, data_source_id)
+      @names = names
+      @batch_data = batch_data
+      @data_source_id = data_source_id
+      @stats = Stats.new
+    end
+
+    def run
+      res = remote_resolve(@names)
+      [res, @batch_data, @stats]
+    end
+
+    private
+
+    def variables(names)
+      { dataSourceIds: [@data_source_id],
+        names: names.
+          map { |name| { value: name[:name], suppliedId: name[:id] } } }
+    end
+
+    def remote_resolve(names)
+      batch_start = Time.now
+      res = GRAPHQL.client.query(QUERY, variables: variables(names))
+      stats_add_batch_time(batch_start)
+      res.data.name_resolver.responses
+    end
+
+    def stats_add_batch_time(batch_start)
+      @stats.stats[:last_batches_time] << Time.now - batch_start
+    end
+  end
+end