+++ spec/lib/gn_list_resolver_spec.rb
@@ -81,8 +81,9 @@ describe GnListResolver do
         subject.run(opts5) do |stats|
           states << stats[:status]
           expect(stats[:total_records]).to be 301
-          expect([0, 301].include?(stats[:resolved_records])).to be true
-          matches = stats[:matches].values.inject(0, :+)
+          expect([0, 101, 201, 200, 301].include?(stats[:resolved_records])).
+            to be true
+          matches = stats[:matches].values.inject(:+)
           expect(matches).to be stats[:resolved_records]
           expect(stats.keys).
             to match_array %i[status total_records ingested_records